requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
aiohttp=3.9.3=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
apex=0.1=pypi_0
appdirs=1.4.4=pyh9f0ad1d_0
arrow-cpp=14.0.2=h374c478_1
astunparse=1.6.3=pypi_0
attrs=23.1.0=py311h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.7=py311hf4808d0_0
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py311h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
cuda=11.8.0=0
cuda-cccl=11.8.89=0
cuda-command-line-tools=11.8.0=0
cuda-compiler=11.8.0=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cuobjdump=11.8.86=0
cuda-cupti=11.8.87=0
cuda-cuxxfilt=11.8.86=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.8.89=0
cuda-gdb=11.8.86=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.8.89=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.8.86=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.8.86=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-nvvp=11.8.87=0
cuda-profiler-api=11.8.86=0
cuda-runtime=11.8.0=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.8.0=0
cuda-tools=11.8.0=0
cuda-visual-tools=11.8.0=0
datasets=2.12.0=py311h06a4308_0
dill=0.3.6=py311h06a4308_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=py_0
fake-factory=9999.9.9=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py311h5eee18b_0
fsspec=2023.9.2=pypi_0
gast=0.5.4=pypi_0
gds-tools=1.4.0.31=0
gflags=2.2.2=h6a678d5_1
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.42=pyhd8ed1ab_0
glog=0.5.0=h6a678d5_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
huggingface_hub=0.20.3=py311h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py311h06a4308_0
importlib-metadata=7.0.1=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py311h06a4308_0
jpeg=9e=h5eee18b_1
kaggle=1.6.11=pyhd8ed1ab_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.9.0.20=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.5.119=0
libcurand-dev=10.3.0.86=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py311h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.4=py311h5eee18b_0
multiprocess=0.70.14=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
nvidia-dali-cuda110=1.36.0=pypi_0
nvidia-nvimgcodec-cu11=0.2.0.7=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_1
orc=1.7.4=hb3bc3d3_1
packaging=23.2=py311h06a4308_0
pandas=2.2.1=py311ha02d727_0
pathtools=0.1.2=py_1
pillow=10.2.0=py311h5eee18b_0
pip=23.3.1=py311h06a4308_0
protobuf=3.20.3=py311hcafe171_1
psutil=5.9.8=py311h459d7ec_0
pyarrow=14.0.2=py311hb6e97c4_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.8=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-utils=3.8.2=pypi_0
python-xxhash=2.0.2=py311h5eee18b_1
python_abi=3.11=2_cp311
pytorch=2.2.1=py3.11_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py311h06a4308_0
pyyaml=6.0.1=py311h5eee18b_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py311h5eee18b_0
requests=2.31.0=py311h06a4308_1
responses=0.13.3=pyhd3eb1b0_0
s2n=1.3.27=hdbd6064_0
safetensors=0.4.2=py311h24d97f6_0
scipy=1.12.0=py311h08b1b3b_0
sentry-sdk=1.43.0=pyhd8ed1ab_0
setproctitle=1.3.3=py311h459d7ec_0
setuptools=68.2.2=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.1.10=h6a678d5_1
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
text-unidecode=1.3=pyhd3eb1b0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=py311h22610ee_0
torch-utils=0.1.2=pypi_0
torchaudio=2.2.1=py311_cu118
torchtriton=2.2.0=py311
torchvision=0.17.1=py311_cu118
tqdm=4.65.0=py311h92b7b1e_0
transformers=4.37.2=py311h06a4308_0
typing-extensions=4.9.0=py311h06a4308_1
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2024a=h04d1e81_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=2.1.0=py311h06a4308_1
utf8proc=2.6.1=h5eee18b_1
wandb=0.16.3=pyhd8ed1ab_0
webencodings=0.5.1=py311h06a4308_1
wheel=0.41.2=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py311h5eee18b_0
zipp=3.17.0=py311h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
